# docs/swagger/swagger.yaml
definitions:
  domain.GatewayApplication:
    properties:
      cluster:
        type: string
      gatewayId:
        type: string
      sparkApplication:
        $ref: '#/definitions/domain.GatewaySparkApplication'
      sparkLogURLs:
        $ref: '#/definitions/domain.SparkLogURLs'
      user:
        type: string
    type: object
...
  domain.GatewayApplicationStatus:
    properties:
      applicationState:
        allOf:
        - $ref: '#/definitions/v1beta2.ApplicationState'
        description: AppState tells the overall application state.
      driverInfo:
        allOf:
        - $ref: '#/definitions/v1beta2.DriverInfo'
        description: DriverInfo has information about the driver.
      executionAttempts:
        description: |-
          ExecutionAttempts is the total number of attempts to run a submitted application to completion.
          Incremented upon each attempted run of the application and reset upon invalidation.
        type: integer
      executorState:
        additionalProperties:
          $ref: '#/definitions/v1beta2.ExecutorState'
        description: ExecutorState records the state of executors by executor Pod
          names.
        type: object
      lastSubmissionAttemptTime:
        description: |-
          LastSubmissionAttemptTime is the time for the last application submission attempt.
          +nullable
        type: string
      sparkApplicationId:
        description: SparkApplicationID is set by the spark-distribution(via spark.app.id
          config) on the driver and executor pods
        type: string
      submissionAttempts:
        description: |-
          SubmissionAttempts is the total number of attempts to submit an application to run.
          Incremented upon each attempted submission of the application and reset upon invalidation and rerun.
        type: integer
      submissionID:
        description: SubmissionID is a unique ID of the current submission of the
          application.
        type: string
      terminationTime:
        description: |-
          CompletionTime is the time when the application runs to completion if it does.
          +nullable
        type: string
    type: object
  domain.GatewayApplicationSummary:
    properties:
      annotations:
        additionalProperties:
          type: string
        type: object
      apiVersion:
        type: string
      cluster:
        type: string
      gatewayId:
        type: string
      kind:
        type: string
      labels:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
      namespace:
        type: string
      status:
        $ref: '#/definitions/domain.GatewayApplicationStatus'
    type: object
  domain.GatewaySparkApplication:
    properties:
      metadata:
        $ref: '#/definitions/domain.GatewayApplicationMeta'
      spec:
        $ref: '#/definitions/domain.GatewayApplicationSpec'
      status:
        $ref: '#/definitions/domain.GatewayApplicationStatus'
    type: object
  domain.SparkLogURLs:
    properties:
      logsUI:
        type: string
      sparkHistoryUI:
        type: string
      sparkUI:
        type: string
    type: object
...
  v1.ObjectMeta:
    properties:
      annotations:
        additionalProperties:
          type: string
        description: |-
          Annotations is an unstructured key value map stored with a resource that may be
          set by external tools to store and retrieve arbitrary metadata. They are not
          queryable and should be preserved when modifying objects.
          More info: https://kubernetes.io/docs/concepts/overview/working-with-objects/annotations
          +optional
        type: object
      creationTimestamp:
        description: |-
          CreationTimestamp is a timestamp representing the server time when this object was
          created. It is not guaranteed to be set in happens-before order across separate operations.
          Clients may not set this value. It is represented in RFC3339 form and is in UTC.

          Populated by the system.
          Read-only.
          Null for lists.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata
          +optional
        type: string
      deletionGracePeriodSeconds:
        description: |-
          Number of seconds allowed for this object to gracefully terminate before
          it will be removed from the system. Only set when deletionTimestamp is also set.
          May only be shortened.
          Read-only.
          +optional
        type: integer
      deletionTimestamp:
        description: |-
          DeletionTimestamp is RFC 3339 date and time at which this resource will be deleted. This
          field is set by the server when a graceful deletion is requested by the user, and is not
          directly settable by a client. The resource is expected to be deleted (no longer visible
          from resource lists, and not reachable by name) after the time in this field, once the
          finalizers list is empty. As long as the finalizers list contains items, deletion is blocked.
          Once the deletionTimestamp is set, this value may not be unset or be set further into the
          future, although it may be shortened or the resource may be deleted prior to this time.
          For example, a user may request that a pod is deleted in 30 seconds. The Kubelet will react
          by sending a graceful termination signal to the containers in the pod. After that 30 seconds,
          the Kubelet will send a hard termination signal (SIGKILL) to the container and after cleanup,
          remove the pod from the API. In the presence of network partitions, this object may still
          exist after this timestamp, until an administrator or automated process can determine the
          resource is fully terminated.
          If not set, graceful deletion of the object has not been requested.

          Populated by the system when a graceful deletion is requested.
          Read-only.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata
          +optional
        type: string
      finalizers:
        description: |-
          Must be empty before the object is deleted from the registry. Each entry
          is an identifier for the responsible component that will remove the entry
          from the list. If the deletionTimestamp of the object is non-nil, entries
          in this list can only be removed.
          Finalizers may be processed and removed in any order.  Order is NOT enforced
          because it introduces significant risk of stuck finalizers.
          finalizers is a shared field, any actor with permission can reorder it.
          If the finalizer list is processed in order, then this can lead to a situation
          in which the component responsible for the first finalizer in the list is
          waiting for a signal (field value, external system, or other) produced by a
          component responsible for a finalizer later in the list, resulting in a deadlock.
          Without enforced ordering finalizers are free to order amongst themselves and
          are not vulnerable to ordering changes in the list.
          +optional
          +patchStrategy=merge
          +listType=set
        items:
          type: string
        type: array
      generateName:
        description: |-
          GenerateName is an optional prefix, used by the server, to generate a unique
          name ONLY IF the Name field has not been provided.
          If this field is used, the name returned to the client will be different
          than the name passed. This value will also be combined with a unique suffix.
          The provided value has the same validation rules as the Name field,
          and may be truncated by the length of the suffix required to make the value
          unique on the server.

          If this field is specified and the generated name exists, the server will return a 409.

          Applied only if Name is not specified.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#idempotency
          +optional
        type: string
      generation:
        description: |-
          A sequence number representing a specific generation of the desired state.
          Populated by the system. Read-only.
          +optional
        type: integer
      labels:
        additionalProperties:
          type: string
        description: |-
          Map of string keys and values that can be used to organize and categorize
          (scope and select) objects. May match selectors of replication controllers
          and services.
          More info: https://kubernetes.io/docs/concepts/overview/working-with-objects/labels
          +optional
        type: object
      managedFields:
        description: |-
          ManagedFields maps workflow-id and version to the set of fields
          that are managed by that workflow. This is mostly for internal
          housekeeping, and users typically shouldn't need to set or
          understand this field. A workflow can be the user's name, a
          controller's name, or the name of a specific apply path like
          "ci-cd". The set of fields is always in the version that the
          workflow used when modifying the object.

          +optional
          +listType=atomic
        items:
          $ref: '#/definitions/v1.ManagedFieldsEntry'
        type: array
      name:
        description: |-
          Name must be unique within a namespace. Is required when creating resources, although
          some resources may allow a client to request the generation of an appropriate name
          automatically. Name is primarily intended for creation idempotence and configuration
          definition.
          Cannot be updated.
          More info: https://kubernetes.io/docs/concepts/overview/working-with-objects/names#names
          +optional
        type: string
      namespace:
        description: |-
          Namespace defines the space within which each name must be unique. An empty namespace is
          equivalent to the "default" namespace, but "default" is the canonical representation.
          Not all objects are required to be scoped to a namespace - the value of this field for
          those objects will be empty.

          Must be a DNS_LABEL.
          Cannot be updated.
          More info: https://kubernetes.io/docs/concepts/overview/working-with-objects/namespaces
          +optional
        type: string
      ownerReferences:
        description: |-
          List of objects depended by this object. If ALL objects in the list have
          been deleted, this object will be garbage collected. If this object is managed by a controller,
          then an entry in this list will point to this controller, with the controller field set to true.
          There cannot be more than one managing controller.
          +optional
          +patchMergeKey=uid
          +patchStrategy=merge
          +listType=map
          +listMapKey=uid
        items:
          $ref: '#/definitions/v1.OwnerReference'
        type: array
      resourceVersion:
        description: |-
          An opaque value that represents the internal version of this object that can
          be used by clients to determine when objects have changed. May be used for optimistic
          concurrency, change detection, and the watch operation on a resource or set of resources.
          Clients must treat these values as opaque and passed unmodified back to the server.
          They may only be valid for a particular resource or set of resources.

          Populated by the system.
          Read-only.
          Value must be treated as opaque by clients and .
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#concurrency-control-and-consistency
          +optional
        type: string
      selfLink:
        description: |-
          Deprecated: selfLink is a legacy read-only field that is no longer populated by the system.
          +optional
        type: string
      uid:
        description: |-
          UID is the unique in time and space value for this object. It is typically generated by
          the server on successful creation of a resource and is not allowed to change on PUT
          operations.

          Populated by the system.
          Read-only.
          More info: https://kubernetes.io/docs/concepts/overview/working-with-objects/names#uids
          +optional
        type: string
    type: object
...
  v1beta2.ApplicationState:
    properties:
      errorMessage:
        type: string
      state:
        $ref: '#/definitions/v1beta2.ApplicationStateType'
    type: object
...
  v1beta2.DeployMode:
    enum:
    - cluster
    - client
    - in-cluster-client
    type: string
    x-enum-varnames:
    - DeployModeCluster
    - DeployModeClient
    - DeployModeInClusterClient
  v1beta2.DriverInfo:
    properties:
      podName:
        type: string
      webUIAddress:
        description: UI Details for the UI created via ClusterIP service accessible
          from within the cluster.
        type: string
      webUIIngressAddress:
        type: string
      webUIIngressName:
        description: Ingress Details if an ingress for the UI was created.
        type: string
      webUIPort:
        type: integer
      webUIServiceName:
        type: string
    type: object
...
  v1beta2.ExecutorState:
    enum:
    - PENDING
    - RUNNING
    - COMPLETED
    - FAILED
    - UNKNOWN
    type: string
    x-enum-varnames:
    - ExecutorStatePending
    - ExecutorStateRunning
    - ExecutorStateCompleted
    - ExecutorStateFailed
    - ExecutorStateUnknown
...
  v1beta2.SparkApplication:
    properties:
      apiVersion:
        description: |-
          APIVersion defines the versioned schema of this representation of an object.
          Servers should convert recognized schemas to the latest internal value, and
          may reject unrecognized values.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources
          +optional
        type: string
      kind:
        description: |-
          Kind is a string value representing the REST resource this object represents.
          Servers may infer this from the endpoint the client submits requests to.
          Cannot be updated.
          In CamelCase.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds
          +optional
        type: string
      metadata:
        $ref: '#/definitions/v1.ObjectMeta'
      spec:
        $ref: '#/definitions/v1beta2.SparkApplicationSpec'
      status:
        $ref: '#/definitions/v1beta2.SparkApplicationStatus'
    type: object
  v1beta2.SparkApplicationSpec:
    properties:
      arguments:
        description: |-
          Arguments is a list of arguments to be passed to the application.
          +optional
        items:
          type: string
        type: array
      batchScheduler:
        description: |-
          BatchScheduler configures which batch scheduler will be used for scheduling
          +optional
        type: string
      batchSchedulerOptions:
        allOf:
        - $ref: '#/definitions/v1beta2.BatchSchedulerConfiguration'
        description: |-
          BatchSchedulerOptions provides fine-grained control on how to batch scheduling.
          +optional
      deps:
        allOf:
        - $ref: '#/definitions/v1beta2.Dependencies'
        description: |-
          Deps captures all possible types of dependencies of a Spark application.
          +optional
      driver:
        allOf:
        - $ref: '#/definitions/v1beta2.DriverSpec'
        description: Driver is the driver specification.
      driverIngressOptions:
        description: |-
          DriverIngressOptions allows configuring the Service and the Ingress to expose ports inside Spark Driver
          +optional
        items:
          $ref: '#/definitions/v1beta2.DriverIngressConfiguration'
        type: array
      dynamicAllocation:
        allOf:
        - $ref: '#/definitions/v1beta2.DynamicAllocation'
        description: |-
          DynamicAllocation configures dynamic allocation that becomes available for the Kubernetes
          scheduler backend since Spark 3.0.
          +optional
      executor:
        allOf:
        - $ref: '#/definitions/v1beta2.ExecutorSpec'
        description: Executor is the executor specification.
      failureRetries:
        description: |-
          FailureRetries is the number of times to retry a failed application before giving up.
          This is best effort and actual retry attempts can be >= the value specified.
          +optional
        type: integer
      hadoopConf:
        additionalProperties:
          type: string
        description: |-
          HadoopConf carries user-specified Hadoop configuration properties as they would use the "--conf" option
          in spark-submit. The SparkApplication controller automatically adds prefix "spark.hadoop." to Hadoop
          configuration properties.
          +optional
        type: object
      hadoopConfigMap:
        description: |-
          HadoopConfigMap carries the name of the ConfigMap containing Hadoop configuration files such as core-site.xml.
          The controller will add environment variable HADOOP_CONF_DIR to the path where the ConfigMap is mounted to.
          +optional
        type: string
      image:
        description: |-
          Image is the container image for the driver, executor, and init-container. Any custom container images for the
          driver, executor, or init-container takes precedence over this.
          +optional
        type: string
      imagePullPolicy:
        description: |-
          ImagePullPolicy is the image pull policy for the driver, executor, and init-container.
          +optional
        type: string
      imagePullSecrets:
        description: |-
          ImagePullSecrets is the list of image-pull secrets.
          +optional
        items:
          type: string
        type: array
      mainApplicationFile:
        description: MainFile is the path to a bundled JAR, Python, or R file of the
          application.
        type: string
      mainClass:
        description: |-
          MainClass is the fully-qualified main class of the Spark application.
          This only applies to Java/Scala Spark applications.
          +optional
        type: string
      memoryOverheadFactor:
        description: |-
          This sets the Memory Overhead Factor that will allocate memory to non-JVM memory.
          For JVM-based jobs this value will default to 0.10, for non-JVM jobs 0.40. Value of this field will
          be overridden by `Spec.Driver.MemoryOverhead` and `Spec.Executor.MemoryOverhead` if they are set.
          +optional
        type: string
      mode:
        allOf:
        - $ref: '#/definitions/v1beta2.DeployMode'
        description: |-
          Mode is the deployment mode of the Spark application.
          +kubebuilder:validation:Enum={cluster,client}
      monitoring:
        allOf:
        - $ref: '#/definitions/v1beta2.MonitoringSpec'
        description: |-
          Monitoring configures how monitoring is handled.
          +optional
      nodeSelector:
        additionalProperties:
          type: string
        description: |-
          NodeSelector is the Kubernetes node selector to be added to the driver and executor pods.
          This field is mutually exclusive with nodeSelector at podSpec level (driver or executor).
          This field will be deprecated in future versions (at SparkApplicationSpec level).
          +optional
        type: object
      proxyUser:
        description: |-
          ProxyUser specifies the user to impersonate when submitting the application.
          It maps to the command-line flag "--proxy-user" in spark-submit.
          +optional
        type: string
      pythonVersion:
        description: |-
          This sets the major Python version of the docker
          image used to run the driver and executor containers. Can either be 2 or 3, default 2.
          +optional
          +kubebuilder:validation:Enum={"2","3"}
        type: string
      restartPolicy:
        allOf:
        - $ref: '#/definitions/v1beta2.RestartPolicy'
        description: RestartPolicy defines the policy on if and in which conditions
          the controller should restart an application.
      retryInterval:
        description: |-
          RetryInterval is the unit of intervals in seconds between submission retries.
          +optional
        type: integer
      sparkConf:
        additionalProperties:
          type: string
        description: |-
          SparkConf carries user-specified Spark configuration properties as they would use the  "--conf" option in
          spark-submit.
          +optional
        type: object
      sparkConfigMap:
        description: |-
          SparkConfigMap carries the name of the ConfigMap containing Spark configuration files such as log4j.properties.
          The controller will add environment variable SPARK_CONF_DIR to the path where the ConfigMap is mounted to.
          +optional
        type: string
      sparkUIOptions:
        allOf:
        - $ref: '#/definitions/v1beta2.SparkUIConfiguration'
        description: |-
          SparkUIOptions allows configuring the Service and the Ingress to expose the sparkUI
          +optional
      sparkVersion:
        description: SparkVersion is the version of Spark the application uses.
        type: string
      timeToLiveSeconds:
        description: |-
          TimeToLiveSeconds defines the Time-To-Live (TTL) duration in seconds for this SparkApplication
          after its termination.
          The SparkApplication object will be garbage collected if the current time is more than the
          TimeToLiveSeconds since its termination.
          +optional
        type: integer
      type:
        allOf:
        - $ref: '#/definitions/v1beta2.SparkApplicationType'
        description: |-
          Type tells the type of the Spark application.
          +kubebuilder:validation:Enum={Java,Python,Scala,R}
      volumes:
        description: |-
          Volumes is the list of Kubernetes volumes that can be mounted by the driver and/or executors.
          +optional
        items:
          $ref: '#/definitions/v1.Volume'
        type: array
    type: object
  v1beta2.SparkApplicationStatus:
    properties:
      applicationState:
        allOf:
        - $ref: '#/definitions/v1beta2.ApplicationState'
        description: AppState tells the overall application state.
      driverInfo:
        allOf:
        - $ref: '#/definitions/v1beta2.DriverInfo'
        description: DriverInfo has information about the driver.
      executionAttempts:
        description: |-
          ExecutionAttempts is the total number of attempts to run a submitted application to completion.
          Incremented upon each attempted run of the application and reset upon invalidation.
        type: integer
      executorState:
        additionalProperties:
          $ref: '#/definitions/v1beta2.ExecutorState'
        description: ExecutorState records the state of executors by executor Pod
          names.
        type: object
      lastSubmissionAttemptTime:
        description: |-
          LastSubmissionAttemptTime is the time for the last application submission attempt.
          +nullable
        type: string
      sparkApplicationId:
        description: SparkApplicationID is set by the spark-distribution(via spark.app.id
          config) on the driver and executor pods
        type: string
      submissionAttempts:
        description: |-
          SubmissionAttempts is the total number of attempts to submit an application to run.
          Incremented upon each attempted submission of the application and reset upon invalidation and rerun.
        type: integer
      submissionID:
        description: SubmissionID is a unique ID of the current submission of the
          application.
        type: string
      terminationTime:
        description: |-
          CompletionTime is the time when the application runs to completion if it does.
          +nullable
        type: string
    type: object
  v1beta2.SparkApplicationType:
    enum:
    - Java
    - Scala
    - Python
    - R
    type: string
    x-enum-varnames:
    - SparkApplicationTypeJava
    - SparkApplicationTypeScala
    - SparkApplicationTypePython
    - SparkApplicationTypeR
...
info:
  contact: {}
  description: REST API for managing GatewayApplication resources across multiple
    clusters
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Spark Gateway
  version: "1.0"
paths:
  /v1/applications/:
    get:
      consumes:
      - application/json
      description: Lists summaries of applications in specified cluster. Optionally
        filter by namespace.
      parameters:
      - description: Cluster name
        in: query
        name: cluster
        required: true
        type: string
      - description: Namespace (optional)
        in: query
        name: namespace
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of GatewayApplicationSummary objects
          schema:
            items:
              $ref: '#/definitions/domain.GatewayApplicationSummary'
            type: array
      security:
      - BasicAuth: []
      summary: List GatewayApplicationSummary
      tags:
      - Applications
    post:
      consumes:
      - application/json
      description: Submits the provided GatewayApplication to the given namespace.
      parameters:
      - description: v1beta2.SparkApplication resource
        in: body
        name: SparkApplication
        required: true
        schema:
          $ref: '#/definitions/v1beta2.SparkApplication'
      produces:
      - application/json
      responses:
        "201":
          description: GatewayApplication Created
          schema:
            $ref: '#/definitions/domain.GatewayApplication'
      security:
      - BasicAuth: []
      summary: Submit a new GatewayApplication
      tags:
      - Applications
  /v1/applications/{gatewayId}:
    delete:
      consumes:
      - application/json
      description: Deletes the specified GatewayApplication
      parameters:
      - description: GatewayApplication Name
        in: path
        name: gatewayId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'Application deleted: {''status'': ''success''}'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BasicAuth: []
      summary: Delete a GatewayApplication
      tags:
      - Applications
    get:
      consumes:
      - application/json
      description: Retrieves the full GatewayApplication resource by ID.
      parameters:
      - description: GatewayApplication Name
        in: path
        name: gatewayId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GatewayApplication resource
          schema:
            $ref: '#/definitions/domain.GatewayApplication'
      security:
      - BasicAuth: []
      summary: Get a GatewayApplication
      tags:
      - Applications
  /v1/applications/{gatewayId}/logs:
    get:
      consumes:
      - application/json
      description: Retrieves the last N lines of driver logs for the specified GatewayApplication.
        Defaults to the last 100 lines.
      parameters:
      - description: GatewayApplication Name
        in: path
        name: gatewayId
        required: true
        type: string
      - description: 'Number of log lines to retrieve (default: 100)'
        in: query
        name: lines
        type: integer
      produces:
      - text/plain
      responses:
        "200":
          description: Driver logs
          schema:
            type: string
      security:
      - BasicAuth: []
      summary: Get driver logs of a GatewayApplication
      tags:
      - Applications
  /v1/applications/{gatewayId}/status:
    get:
      consumes:
      - application/json
      description: Retrieves only the status field of a GatewayApplication.
      parameters:
      - description: GatewayApplication Name
        in: path
        name: gatewayId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GatewayApplication status
          schema:
            $ref: '#/definitions/v1beta2.SparkApplicationStatus'
      security:
      - BasicAuth: []
      summary: Get GatewayApplication status
      tags:
      - Applications
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"